# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Progi CAP API",
        "description": "API documentation for Progi CAP application modules",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/files/{file}/download": {
            "get": {
                "tags": [
                    "File Management"
                ],
                "summary": "Télécharger un fichier",
                "description": "Télécharge un fichier spécifique",
                "operationId": "downloadFile",
                "parameters": [
                    {
                        "name": "file",
                        "in": "path",
                        "description": "ID du fichier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier téléchargé avec succès",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Fichier non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/files/{file}/shares": {
            "get": {
                "tags": [
                    "File Sharing"
                ],
                "summary": "Partages d'un fichier",
                "description": "Récupère tous les liens de partage d'un fichier",
                "operationId": "getFileShares",
                "parameters": [
                    {
                        "name": "file",
                        "in": "path",
                        "description": "ID du fichier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partages récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FileShare"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Fichier non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "File Sharing"
                ],
                "summary": "Créer un partage",
                "description": "Crée un nouveau lien de partage pour un fichier",
                "operationId": "createFileShare",
                "parameters": [
                    {
                        "name": "file",
                        "in": "path",
                        "description": "ID du fichier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "description": "Mot de passe pour accéder au partage",
                                        "type": "string"
                                    },
                                    "allow_download": {
                                        "description": "Autoriser le téléchargement",
                                        "type": "boolean",
                                        "default": true
                                    },
                                    "allow_preview": {
                                        "description": "Autoriser l'aperçu",
                                        "type": "boolean",
                                        "default": true
                                    },
                                    "max_downloads": {
                                        "description": "Nombre maximum de téléchargements",
                                        "type": "integer"
                                    },
                                    "expires_at": {
                                        "description": "Date d'expiration",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Partage créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lien de partage créé avec succès."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/FileShare"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Fichier non trouvé"
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/files/{file}/shares/{share}": {
            "get": {
                "tags": [
                    "File Sharing"
                ],
                "summary": "Détails d'un partage",
                "description": "Affiche les détails d'un lien de partage spécifique",
                "operationId": "getFileShare",
                "parameters": [
                    {
                        "name": "file",
                        "in": "path",
                        "description": "ID du fichier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "share",
                        "in": "path",
                        "description": "ID du partage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du partage récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/FileShare"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Fichier ou partage non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "File Sharing"
                ],
                "summary": "Supprimer un partage",
                "description": "Supprime définitivement un lien de partage",
                "operationId": "deleteFileShare",
                "parameters": [
                    {
                        "name": "file",
                        "in": "path",
                        "description": "ID du fichier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "share",
                        "in": "path",
                        "description": "ID du partage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partage supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Partage supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Fichier ou partage non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/files/{file}/shares/{share}/deactivate": {
            "post": {
                "tags": [
                    "File Sharing"
                ],
                "summary": "Désactiver un partage",
                "description": "Désactive un lien de partage existant",
                "operationId": "deactivateFileShare",
                "parameters": [
                    {
                        "name": "file",
                        "in": "path",
                        "description": "ID du fichier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "share",
                        "in": "path",
                        "description": "ID du partage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partage désactivé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Partage désactivé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Fichier ou partage non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}